% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banerjee.R
\name{banerjee}
\alias{banerjee}
\title{Banerjee Test}
\usage{
banerjee(formula, data, lags = 1, trend = "const")
}
\arguments{
\item{formula}{An object of class "formula" to describe the model.}

\item{data}{An optional data frame containing the variables in the model.}

\item{lags}{Number of lags to be included.}

\item{trend}{Type of deterministic component to be inlcuded, "none" for no deterministics,
"const" for a constant and "trend" for a constant plus trend.}
}
\value{
\code{banerjee} returns an object of class \code{"co.test"}.
}
\description{
Executes Banerjee Test.
}
\examples{
banerjee(linvestment ~ lincome + lconsumption, data = Lutkepohl)
}
\references{
Banerjee, A., Dolado, J. J. and Mestre, R. (1998), Error-correction Mechanism Tests for Cointegration in a Single-equation Framework, Journal of Times Series Analysis 19(3), 267-283.
}
