% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boswijk.R
\name{boswijk}
\alias{boswijk}
\title{Boswijk Test}
\usage{
boswijk(formula, data, lags = 1, trend = "const")
}
\arguments{
\item{formula}{An object of class "formula" to describe the model.}

\item{data}{An optional data frame containing the variables in the model.}

\item{lags}{Number of lags to be included.}

\item{trend}{Type of deterministic component to be inlcuded, "none" for no deterministics,
"const" for a constant and "trend" for a constant plus trend.}
}
\value{
\code{boswijk} returns an object of class \code{"co.test"}.
}
\description{
Executes Boswijk Test.
}
\examples{
boswijk(linvestment ~ lincome + lconsumption, data = Lutkepohl)
}
\references{
Boswijk, H. P. (1994), Testing for an unstable root in conditional and structural error correction models, Journal of Econometrics 63(1), 37-60.
}
