% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/englegranger.R
\name{englegranger}
\alias{englegranger}
\title{Engle-Granger Test}
\usage{
englegranger(formula, data, lags = 1, trend = "const")
}
\arguments{
\item{formula}{An object of class "formula" to describe the model.}

\item{data}{An optional data frame containing the variables in the model.}

\item{lags}{Number of lags to be included.}

\item{trend}{Type of deterministic component to be inlcuded, "none" for no deterministics,
"const" for a constant and "trend" for a constant plus trend.}
}
\value{

}
\description{
Execute Engle-Granger Test.
}
