% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.co.test}
\alias{summary.co.test}
\title{Summary method for objects from bayerhanck}
\usage{
## S3 method for class "co.test"
## S3 method for class "bayerhanck"
}
\arguments{
\item{object}{An object of class \code{co.test} or \code{bayerhanck}.}
}
\value{

}
\description{
Summary method for objects from bayerhanck
}
