% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayerhanck.R, R/bayerhanck_1.R
\name{bayerhanck}
\alias{bayerhanck}
\alias{bayerhanck_1}
\title{Joint Test-Statistic for the Null of Non-Cointegration}
\usage{
bayerhanck(formula, data, lags = 1, trend = "const", test = "all", crit = 0.05)

bayerhanck_1(formula, data, lags = 1, trend = "const", test = "all", crit = 0.05)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} to describe the model.}

\item{data}{An optional data frame containing the variables in the model.}

\item{lags}{Number of lags to be included.}

\item{trend}{Type of deterministic component to be included, "none" for no deterministic,
"const" for a constant and "trend" for a constant plus trend.}

\item{test}{Selection of tests to choose from. Choices are either "eg-j", for \code{\link{englegranger}}
and \code{\link{johansen}}, or "all", for \code{\link{englegranger}}, \code{\link{johansen}},
\code{\link{banerjee}} and \code{\link{boswijk}}.}

\item{crit}{Level for the critical value of the test to be reported.}
}
\value{
\code{bayerhanck} returns an object of classes \code{"bh.test"} and \code{"list"}.

The function \code{summary} is used to print a summary, whereas the cumulative distribution
under the null hypothesis can be plotted with \code{plot}.
}
\description{
Produces a joint test-statistic for the null of non-cointegration, aggregating
various cointegration tests.
}
\section{Functions}{
\itemize{
\item \code{bayerhanck}: Bayer-Hanck Test-Statistic with values obtained from the technical appendix of Bayer and Hanck (2009).

\item \code{bayerhanck_1}: Bayer-Hanck Test-Statistic with values obtained by own simulation.
}}

\examples{
data("mts-examples", package="MTS")
bayerhanck(sp ~ ibm + ko, data = ibmspko)
}
\references{
Bayer, C. and Hanck, C. (2009), Combining Non-Cointegration tests, METEOR RM 09/012, University of Maastricht.
}
